Separate domain model for Twitter entities
