+++ src/ObtainingUserDetails.test.ts
@@ -1,10 +1,7 @@
 import { mock, MockProxy } from "jest-mock-extended";
 
-import {
-  obtainingUserDetails,
-  UserDetails,
-  UserProfileView,
-} from "./ObtainingUserDetails";
+import { obtainingUserDetails, UserProfileView } from "./ObtainingUserDetails";
+import { UserDetails } from "./twitter-model/UserDetails";
 
 describe("Obtaining user details", () => {
   let testProfileView: MockProxy<UserProfileView>;
