+++ src/twitter-model/Tweet.ts
@@ -1,0 +1,6 @@
+export type Tweet = {
+  author: string;
+  url: string;
+  time: Date;
+  text: string;
+};
