+++ src/twitter-model/UserDetails.ts
@@ -1,0 +1,5 @@
+export type UserDetails = {
+  name: string;
+  username: string;
+  bio?: string;
+};
